refactor(github): collapse IPC handlers into thin service orchestrators
- Move buildGitHubSearchQuery from handlers to GitHubQueries.ts
- Extract /rate_limit fetch into new GitHubRateLimitApi.ts service
- Extract token set/clear orchestration into new GitHubTokenOrchestrator.ts
- Extract remote discovery into new GitHubRemotes.ts service
- Add getFirstPageCache and getRepoStatsComplete to GitHubStats.ts
- Export PRIMARY_RESET_BUFFER_MS from GitHubRateLimitService.ts
- Route all handler dynamic imports through services/github/index.js barrel
- github.ts: 796 → 488 lines, each handler now validate → call → return
